requirements.txt
# Core
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1

# YouTube
google-api-python-client==2.131.0

# Video download (YouTube videos + Shorts → direct playback in Telegram)
yt-dlp==2024.12.23

# Web server (Render keep-alive)
flask==3.0.3
